refactor: split position safety and migrate encoder path to TIM2
Extract position safety evaluation into a dedicated module while keeping emergency stop and output cut-off in the control loop.

Migrate the encoder path from TIM4 to TIM2, promote raw counter handling to 32-bit, and update runtime CSV/diagnostic logging to match the new timer path.

Sync README, module docs, steering portal, lifecycle docs, and change logs with the current runtime structure and bench configuration.

+++ Core/Inc/encoder_reader.h
@@ -31,9 +31,9 @@ int32_t EncoderReader_GetCount(void);
 
 /**
  * @brief Get hardware timer raw counter value
- * @return Raw timer counter (0~65535)
+ * @return Raw timer counter
  */
-uint16_t EncoderReader_GetRawCounter(void);
+uint32_t EncoderReader_GetRawCounter(void);
 
 /**
  * @brief Reset encoder count to zero
